refactor: Flatten Golang packages
