+++ pkg/clair/client_test.go
