+++ pkg/clair/model.go
