+++ pkg/http/api/base_handler_test.go
@@ -1,7 +1,7 @@
 package api
 
 import (
-	"github.com/goharbor/harbor-scanner-clair/pkg/model/harbor"
+	"github.com/goharbor/harbor-scanner-clair/pkg/harbor"
 	"github.com/stretchr/testify/assert"
 	"net/http"
 	"net/http/httptest"
