+++ pkg/http/api/v1/handler.go
@@ -4,9 +4,9 @@ import (
 	"encoding/json"
 	"fmt"
 	"github.com/goharbor/harbor-scanner-clair/pkg/etc"
+	"github.com/goharbor/harbor-scanner-clair/pkg/harbor"
 	"github.com/goharbor/harbor-scanner-clair/pkg/http/api"
 	"github.com/goharbor/harbor-scanner-clair/pkg/job"
-	"github.com/goharbor/harbor-scanner-clair/pkg/model/harbor"
 	"github.com/goharbor/harbor-scanner-clair/pkg/persistence"
 	"github.com/goharbor/harbor-scanner-clair/pkg/scanner"
 	"github.com/gorilla/mux"
