+++ pkg/mock/expectation.go
@@ -18,11 +18,6 @@ func ApplyExpectations(t *testing.T, mock interface{}, expectations ...*Expectat
 		return
 	}
 	switch v := mock.(type) {
-	case *Scanner:
-		m := mock.(*Scanner)
-		for _, e := range expectations {
-			m.On(e.Method, e.Args...).Return(e.ReturnArgs...)
-		}
 	case *ClairClient:
 		m := mock.(*ClairClient)
 		for _, e := range expectations {
