+++ pkg/mock/registry_client.go
@@ -2,7 +2,7 @@ package mock
 
 import (
 	"github.com/docker/distribution"
-	"github.com/goharbor/harbor-scanner-clair/pkg/model/harbor"
+	"github.com/goharbor/harbor-scanner-clair/pkg/harbor"
 	"github.com/goharbor/harbor-scanner-clair/pkg/registry"
 	"github.com/stretchr/testify/mock"
 )
