+++ pkg/mock/store.go
@@ -1,8 +1,8 @@
 package mock
 
 import (
+	"github.com/goharbor/harbor-scanner-clair/pkg/harbor"
 	"github.com/goharbor/harbor-scanner-clair/pkg/job"
-	"github.com/goharbor/harbor-scanner-clair/pkg/model/harbor"
 	"github.com/stretchr/testify/mock"
 )
