+++ pkg/persistence/store.go
@@ -1,8 +1,8 @@
 package persistence
 
 import (
+	"github.com/goharbor/harbor-scanner-clair/pkg/harbor"
 	"github.com/goharbor/harbor-scanner-clair/pkg/job"
-	"github.com/goharbor/harbor-scanner-clair/pkg/model/harbor"
 )
 
 // Store defines methods for persisting ScanJobs and associated ScanReports.
