+++ pkg/scanner/enqueuer.go
@@ -1,10 +1,9 @@
 package scanner
 
 import (
+	"github.com/goharbor/harbor-scanner-clair/pkg/harbor"
 	"github.com/goharbor/harbor-scanner-clair/pkg/job"
-	"github.com/goharbor/harbor-scanner-clair/pkg/model/harbor"
 	"github.com/goharbor/harbor-scanner-clair/pkg/persistence"
-	"github.com/goharbor/harbor-scanner-clair/pkg/scanner/clair"
 	"github.com/goharbor/harbor-scanner-clair/pkg/work"
 	"github.com/google/uuid"
 	"golang.org/x/xerrors"
@@ -17,10 +16,10 @@ type Enqueuer interface {
 }
 
 // NewEnqueuer constructs the default Enqueuer.
-func NewEnqueuer(pool *work.Pool, scanner clair.Scanner, store persistence.Store) Enqueuer {
+func NewEnqueuer(pool *work.Pool, adapter Adapter, store persistence.Store) Enqueuer {
 	return &enqueuer{
 		pool:    pool,
-		scanner: scanner,
+		adapter: adapter,
 		store:   store,
 	}
 }
@@ -28,7 +27,7 @@ func NewEnqueuer(pool *work.Pool, scanner clair.Scanner, store persistence.Store
 type enqueuer struct {
 	store   persistence.Store
 	pool    *work.Pool
-	scanner clair.Scanner
+	adapter Adapter
 }
 
 func (e *enqueuer) Enqueue(request harbor.ScanRequest) (string, error) {
@@ -40,6 +39,11 @@ func (e *enqueuer) Enqueue(request harbor.ScanRequest) (string, error) {
 	if err != nil {
 		return "", xerrors.Errorf("saving scan job: %w", err)
 	}
-	e.pool.Run(NewWorker(e.store, e.scanner, jobID, request))
+	e.pool.Run(&worker{
+		store:   e.store,
+		adapter: e.adapter,
+		jobID:   jobID,
+		request: request,
+	})
 	return jobID, nil
 }
